Split Jasmine into its own project.
