+++ all/pom.xml
@@ -1,0 +1,27 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>org.netmelody.menodora</groupId>
+        <artifactId>menodora-parent</artifactId>
+        <version>0.0.7-SNAPSHOT</version>
+    </parent>
+
+    <artifactId>menodora</artifactId>
+
+    <dependencies>
+        <dependency>
+            <groupId>org.netmelody.menodora</groupId>
+            <artifactId>menodora-core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>org.netmelody.menodora</groupId>
+            <artifactId>menodora-jasmine</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+    </dependencies>
+</project>
