+++ core/src/main/java/org/netmelody/menodora/JavaScriptContext.java
