+++ core/src/main/java/org/netmelody/menodora/core/ExecutionEnvironmentPreparation.java
