+++ core/src/main/java/org/netmelody/menodora/core/JavaScriptEnvironment.java
