+++ core/src/main/java/org/netmelody/menodora/core/JavaScriptReporter.java
