+++ core/src/main/java/org/netmelody/menodora/core/TestStyle.java
