+++ core/src/main/java/org/netmelody/menodora/core/Timer.java
