+++ core/src/main/java/org/netmelody/menodora/core/locator/CompositeLocator.java
