+++ core/src/main/js/env.js-1.2/env.rhino.1.2.js
