+++ core/src/test/java/org/netmelody/menodora/core/ContextTest.java
