+++ core/src/test/resources/dummy-project.jar
