+++ jasmine/src/main/js/jasmine-1.0.2/MIT.LICENSE
