+++ testing/src/main/java/org/netmelody/menodora/testing/DescriptionMatcher.java
@@ -1,4 +1,4 @@
-package org.netmelody.menodora.test;
+package org.netmelody.menodora.testing;
 
 import java.util.Collections;
 import java.util.List;
